+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# Unreleased
+# 0.2.0
+
+* SSL traffic encryption terminating at the ALB
+* DNS resolution
+* Update README
+* Add ALB DNS name as a Terraform output
+* Parameterize ECS CPU and memory
+* ECS autoscaling configuration
 
 # 0.1.0